basic cli features via commander and added support for named profiles
