+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "aws-sdk": "^2.596.0",
     "axios": "^0.19.0",
+    "commander": "^4.1.0",
     "open": "^7.0.0"
   },
   "repository": {
